pyproject.toml
[project]
name = "robo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "catkin-pkg>=1.0.0",
    "empy==3.3.4",
    "hidapi>=0.14.0.post4",
    "lark>=1.2.2",
    "pygame>=2.6.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
]

[tool.ruff]
target-version = "py313"
select = ["ALL"]
fixable = ["ALL"]
ignore = [
    "ERA001",
]

